+++ institutions-sender/src/main/java/com/hiberus/adoptionskafka/controlers/impl/InstitutionControlerImpl.java
@@ -1,14 +1,19 @@
 package com.hiberus.adoptionskafka.controlers.impl;
 
+import com.hiberus.adoptionskafka.Exceptions.AdoptionsMicroUnavailable;
+import com.hiberus.adoptionskafka.Exceptions.InstitutionNotFoundException;
 import com.hiberus.adoptionskafka.controlers.InstitutionControler;
 import com.hiberus.adoptionskafka.dto.EventType;
 import com.hiberus.adoptionskafka.dto.InstitutionDto;
+import com.hiberus.adoptionskafka.services.InstFeignService;
 import com.hiberus.adoptionskafka.services.InstitutionService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import com.google.gson.Gson;
 
 import javax.validation.Valid;
 
@@ -21,6 +26,11 @@ public class InstitutionControlerImpl implements InstitutionControler {
     @Autowired
     InstitutionService institutionService;
 
+    @Autowired
+    InstFeignService instFeignService;
+
+    private final Gson gson = new Gson();
+
     @Override
     @PostMapping
     @ResponseStatus(HttpStatus.ACCEPTED)
@@ -32,18 +42,41 @@ public class InstitutionControlerImpl implements InstitutionControler {
 
     @Override
     @PutMapping(value = "/{idInstitution}")
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    public void modifyInstitution(@PathVariable String idInstitution, @RequestBody InstitutionDto institutionDto) {
+    public ResponseEntity<String> modifyInstitution(@PathVariable String idInstitution, @RequestBody InstitutionDto institutionDto) {
         log.info("Receive http petition for update institution");
-        institutionDto.setEventType(EventType.PUT);
-        institutionService.updateInstitution(idInstitution, institutionDto);
+        try {
+            ResponseEntity<InstitutionDto> responseEntity = instFeignService.checkInstitution(idInstitution);
+            if (responseEntity.getStatusCode().is2xxSuccessful()) {
+                institutionDto.setEventType(EventType.PUT);
+                institutionService.updateInstitution(idInstitution, institutionDto);
+                return new ResponseEntity<>(HttpStatus.ACCEPTED);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        catch (InstitutionNotFoundException e) {
+            return new ResponseEntity<>(gson.toJson(e.getMessage()),HttpStatus.NOT_FOUND);
+        } catch ( AdoptionsMicroUnavailable e) {
+            return new ResponseEntity<>(gson.toJson(e.getMessage()), HttpStatus.SERVICE_UNAVAILABLE);
+        }
     }
 
     @Override
     @DeleteMapping(value = "/{idInstitution}")
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    public void deleteInstitution(@PathVariable String idInstitution) {
+    public ResponseEntity<String> deleteInstitution(@PathVariable String idInstitution) {
         log.info("Receive http petition for delete institution");
-        institutionService.deleteInstitution(idInstitution);
+
+        try {
+            ResponseEntity<InstitutionDto> responseEntity = instFeignService.checkInstitution(idInstitution);
+            if (responseEntity.getStatusCode().is2xxSuccessful()) {
+                institutionService.deleteInstitution(idInstitution);
+                return new ResponseEntity<>(HttpStatus.ACCEPTED);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        catch (InstitutionNotFoundException e) {
+            return new ResponseEntity<>(gson.toJson(e.getMessage()),HttpStatus.NOT_FOUND);
+        } catch ( AdoptionsMicroUnavailable e) {
+            return new ResponseEntity<>(gson.toJson(e.getMessage()), HttpStatus.SERVICE_UNAVAILABLE);
+        }
     }
 }
